Uploaded python 3 prior distribution as well as bug fixes for running prior as a post-processing filter.

+++ utils/data_stats.py
@@ -46,7 +46,7 @@ class DataStats:
             new_image[i] = image.eq(i)
 
         if self.num_classes == 2:
-            new_image[1] += image.eq(2)
+            new_image[1] += image.eq(2).float()
 
         return new_image
